refactor: secure storage move auth logic from accounts composable
