+++ src/composables/accounts.ts
@@ -1,6 +1,5 @@
 import { computed, ref } from 'vue';
 import { uniq } from 'lodash-es';
-import { generateMnemonic, mnemonicToSeed, validateMnemonic } from '@aeternity/bip39';
 
 import type {
   AccountAddress,
@@ -15,178 +14,138 @@ import {
   PROTOCOLS,
   PROTOCOL_LIST,
   STORAGE_KEYS,
-  IS_IOS,
-  IS_MOBILE_APP,
   ACCOUNT_TYPES_LIST,
   ACCOUNT_TYPES,
-  MODAL_PASSWORD_LOGIN,
-  IS_EXTENSION,
-  IS_OFFSCREEN_TAB,
 } from '@/constants';
 import {
-  authenticateWithPassword,
   createCallbackRegistry,
-  encrypt,
-  endSession,
   excludeFalsy,
-  generateEncryptionKey,
-  getSessionEncryptionKey,
   prepareAccountSelectOptions,
-  startSession,
   watchUntilTruthy,
 } from '@/utils';
-import { ProtocolAdapterFactory } from '@/lib/ProtocolAdapterFactory';
 import migrateAccountsVuexToComposable from '@/migrations/001-accounts-vuex-to-composable';
-import migrateMnemonicVuexToComposable from '@/migrations/002-mnemonic-vuex-to-composable';
-import migrateMnemonicCordovaToIonic from '@/migrations/008-mnemonic-cordova-to-ionic';
-import { WalletStorage } from '@/lib/WalletStorage';
-import { SecureMobileStorage } from '@/lib/SecureMobileStorage';
+
+import { ProtocolAdapterFactory } from '@/lib/ProtocolAdapterFactory';
+
 import { useStorageRef } from './storageRef';
-import { useModals } from './modals';
-import { useUi } from './ui';
-import { useSecureStorageRef } from './secureStorageRef';
-import { useAeSdk } from './aeSdk';
-
-const CHECK_FOR_SESSION_KEY_INTERVAL = 5000;
-
-let composableInitialized = false;
+import { createCustomScopedComposable } from './composablesHelpers';
+import { useAuth } from './auth';
 
 const {
   addCallback: onAccountChange,
   runCallbacks: runOnAccountChangeCallbacks,
 } = createCallbackRegistry<(newAccount: IAccount, oldAccount: IAccount) => any>();
 
-const areAccountsRestored = ref(false);
-const isMnemonicRestored = ref(false);
-const encryptionKey = ref<CryptoKey | null>();
-
-const [encryptedMnemonic, mnemonic] = useSecureStorageRef<string>(
-  '',
-  STORAGE_KEYS.mnemonic,
-  {
-    backgroundSync: true,
-    migrations: [
-      ...((IS_IOS && IS_MOBILE_APP) ? [migrateMnemonicCordovaToIonic] : []),
-      migrateMnemonicVuexToComposable,
-    ],
-    onRestored: async (val) => {
-      const hasStoredMnemonic = (
-        WalletStorage.get(STORAGE_KEYS.mnemonic)
-        || await SecureMobileStorage.get(STORAGE_KEYS.mnemonic)
-      );
-      isMnemonicRestored.value = !!val || !hasStoredMnemonic;
-    },
-  },
-);
-
-const accountsRaw = useStorageRef<IAccountRaw[]>(
-  [],
-  STORAGE_KEYS.accountsRaw,
-  {
-    backgroundSync: true,
-    migrations: [
-      migrateAccountsVuexToComposable,
-    ],
-    onRestored: () => {
-      areAccountsRestored.value = true;
-    },
-  },
-);
-
-const activeAccountGlobalIdx = useStorageRef<number>(
-  0,
-  STORAGE_KEYS.activeAccountGlobalIdx,
-  { backgroundSync: true },
-);
-
-const protocolLastActiveGlobalIdx = useStorageRef<ProtocolRecord<number>>(
-  {},
-  STORAGE_KEYS.protocolLastActiveAccountIdx,
-  { backgroundSync: true },
-);
-
-const mnemonicSeed = computed(() => mnemonic.value ? mnemonicToSeed(mnemonic.value) : null);
-
-const accounts = computed((): IAccount[] => {
-  if (!isMnemonicRestored.value || !mnemonic.value || !accountsRaw.value?.length) {
-    return [];
-  }
-
-  // Indexes for each protocol and account type
-  const idxList = PROTOCOL_LIST.reduce(
-    (acc, protocol) => ({
-      ...acc,
-      [protocol]: ACCOUNT_TYPES_LIST.reduce(
-        (acc2, type) => ({ ...acc2, [type]: 0 }),
-        {} as Record<AccountType, number>,
-      ),
-    }),
-    {} as Required<ProtocolRecord<Record<AccountType, number>>>,
-  );
-
-  return accountsRaw.value
-    .map((account, globalIdx) => {
-      const idx = idxList[account.protocol][account.type];
-
-      const adapter = ProtocolAdapterFactory.getAdapter(account.protocol);
-      const resolvedAccount = adapter
-        .resolveAccountRaw(account, idx, globalIdx, mnemonicSeed.value);
-      if (resolvedAccount) {
-        idxList[account.protocol][account.type] += 1;
-      } else {
-        accountsRaw.value.splice(globalIdx, 1);
-      }
-      return resolvedAccount;
-    }).filter(excludeFalsy) as IAccount[];
-});
-
-const activeAccount = computed((): IAccount => accounts.value[activeAccountGlobalIdx.value] || {});
-
-const isActiveAccountAirGap = computed(
-  (): boolean => activeAccount.value.type === ACCOUNT_TYPES.airGap,
-);
-
-const accountsGroupedByProtocol = computed(
-  () => accounts.value.reduce(
-    (acc, account) => ({
-      ...acc,
-      [account.protocol]: [...(acc?.[account.protocol] || []), account],
-    }),
-    {} as ProtocolRecord<IAccount[]>,
-  ),
-);
-
-const aeAccounts = computed(
-  (): IAccount[] => accountsGroupedByProtocol.value[PROTOCOLS.aeternity] || [],
-);
-
-const accountsAddressList = computed(
-  (): string[] => accounts.value.map(({ address }) => address),
-);
-
-const accountsSelectOptions = computed(
-  (): IFormSelectOption[] => prepareAccountSelectOptions(accounts.value),
-);
-
-const aeAccountsSelectOptions = computed(
-  (): IFormSelectOption[] => prepareAccountSelectOptions(aeAccounts.value),
-);
-
-const isLoggedIn = computed(
-  (): boolean => activeAccount.value && Object.keys(activeAccount.value).length > 0,
-);
-
-const protocolsInUse = computed(
-  (): Protocol[] => uniq(accounts.value.map(({ protocol }) => protocol)),
-);
-
 /**
  * Composable that handles all operations related to the accounts.
  * The app is storing only basic account data in the browser storage.
  * The wallets's data is created in fly with the use of computed properties.
  */
-export function useAccounts() {
-  const { secureLoginTimeout, setLoaderVisible } = useUi();
+export const useAccounts = createCustomScopedComposable(() => {
+  const { mnemonic, mnemonicSeed, isMnemonicRestored } = useAuth();
+
+  const areAccountsRestored = ref(false);
+
+  const accountsRaw = useStorageRef<IAccountRaw[]>(
+    [],
+    STORAGE_KEYS.accountsRaw,
+    {
+      backgroundSync: true,
+      migrations: [
+        migrateAccountsVuexToComposable,
+      ],
+      onRestored: () => {
+        areAccountsRestored.value = true;
+      },
+    },
+  );
+
+  const activeAccountGlobalIdx = useStorageRef<number>(
+    0,
+    STORAGE_KEYS.activeAccountGlobalIdx,
+    { backgroundSync: true },
+  );
+
+  const protocolLastActiveGlobalIdx = useStorageRef<ProtocolRecord<number>>(
+    {},
+    STORAGE_KEYS.protocolLastActiveAccountIdx,
+    { backgroundSync: true },
+  );
+
+  const accounts = computed((): IAccount[] => {
+    if (!isMnemonicRestored.value || !mnemonic.value || !accountsRaw.value?.length) {
+      return [];
+    }
+
+    // Indexes for each protocol and account type
+    const idxList = PROTOCOL_LIST.reduce(
+      (acc, protocol) => ({
+        ...acc,
+        [protocol]: ACCOUNT_TYPES_LIST.reduce(
+          (acc2, type) => ({ ...acc2, [type]: 0 }),
+          {} as Record<AccountType, number>,
+        ),
+      }),
+      {} as Required<ProtocolRecord<Record<AccountType, number>>>,
+    );
+
+    return accountsRaw.value
+      .map((account, globalIdx) => {
+        const idx = idxList[account.protocol][account.type];
+
+        const adapter = ProtocolAdapterFactory.getAdapter(account.protocol);
+        const resolvedAccount = adapter
+          .resolveAccountRaw(account, idx, globalIdx, mnemonicSeed.value);
+        if (resolvedAccount) {
+          idxList[account.protocol][account.type] += 1;
+        } else {
+          accountsRaw.value.splice(globalIdx, 1);
+        }
+        return resolvedAccount;
+      }).filter(excludeFalsy) as IAccount[];
+  });
+
+  const activeAccount = computed(
+    (): IAccount => accounts.value[activeAccountGlobalIdx.value] || {},
+  );
+
+  const isActiveAccountAirGap = computed(
+    (): boolean => activeAccount.value.type === ACCOUNT_TYPES.airGap,
+  );
+
+  const accountsGroupedByProtocol = computed(
+    () => accounts.value.reduce(
+      (acc, account) => ({
+        ...acc,
+        [account.protocol]: [...(acc?.[account.protocol] || []), account],
+      }),
+      {} as ProtocolRecord<IAccount[]>,
+    ),
+  );
+
+  const aeAccounts = computed(
+    (): IAccount[] => accountsGroupedByProtocol.value[PROTOCOLS.aeternity] || [],
+  );
+
+  const accountsAddressList = computed(
+    (): string[] => accounts.value.map(({ address }) => address),
+  );
+
+  const accountsSelectOptions = computed(
+    (): IFormSelectOption[] => prepareAccountSelectOptions(accounts.value),
+  );
+
+  const aeAccountsSelectOptions = computed(
+    (): IFormSelectOption[] => prepareAccountSelectOptions(aeAccounts.value),
+  );
+
+  const isLoggedIn = computed(
+    (): boolean => activeAccount.value && Object.keys(activeAccount.value).length > 0,
+  );
+
+  const protocolsInUse = computed(
+    (): Protocol[] => uniq(accounts.value.map(({ protocol }) => protocol)),
+  );
 
   function getAccountByAddress(address: AccountAddress): IAccount | undefined {
     return accounts.value.find((acc) => acc.address === address);
@@ -248,78 +207,6 @@ export function useAccounts() {
   }
 
   /**
-   * Setting/Resetting the password key logs the user in/out.
-   */
-  function setEncryptionKey(newEncryptionKey: CryptoKey | null) {
-    encryptionKey.value = newEncryptionKey;
-    if (IS_EXTENSION) {
-      if (newEncryptionKey) {
-        startSession(newEncryptionKey, secureLoginTimeout.value);
-      } else {
-        endSession();
-      }
-    }
-  }
-
-  async function getEncryptionKey() {
-    return watchUntilTruthy(encryptionKey);
-  }
-
-  async function openPasswordLoginModal() {
-    setLoaderVisible(true);
-    const sessionEncryptionKey = await getSessionEncryptionKey();
-    if (sessionEncryptionKey) {
-      setEncryptionKey(sessionEncryptionKey);
-      const { getAeSdk } = useAeSdk();
-      await getAeSdk();
-      setLoaderVisible(false);
-      return;
-    }
-    setLoaderVisible(false);
-
-    const { openModal } = useModals();
-
-    await openModal(MODAL_PASSWORD_LOGIN);
-    if (!encryptionKey.value) {
-      throw new Error('encryptionKey was not set after login.');
-    }
-  }
-
-  async function setPasswordAndEncryptMnemonic(newMnemonic: string, password: string) {
-    const newEncryptionKey = await generateEncryptionKey(password);
-    const mnemonicEncryptionResult = await encrypt(newEncryptionKey, newMnemonic);
-    encryptedMnemonic.value = mnemonicEncryptionResult;
-    setEncryptionKey(newEncryptionKey);
-  }
-
-  async function updatePassword(currentPassword: string, newPassword: string) {
-    const { decryptedMnemonic } = await authenticateWithPassword(currentPassword);
-    if (decryptedMnemonic) {
-      await setPasswordAndEncryptMnemonic(decryptedMnemonic, newPassword);
-    }
-  }
-
-  async function setMnemonicAndInitializePassword(newMnemonic: string, isRestored = false) {
-    // TODO move the logic related to authentication to `auth` composable
-    if (!IS_MOBILE_APP) {
-      const { openSetPasswordModal } = useModals();
-
-      const password = await openSetPasswordModal(isRestored).catch(() => {
-        throw new Error('Password was not set.');
-      });
-
-      await setPasswordAndEncryptMnemonic(newMnemonic, password);
-    }
-    mnemonic.value = newMnemonic;
-  }
-
-  async function setGeneratedMnemonic() {
-    await setMnemonicAndInitializePassword(generateMnemonic()).catch(() => {
-      throw new Error('Mnemonic was not set.');
-    });
-  }
-
-  /**
    * Determine if provided address belongs to any of the current user's accounts.
    */
   function isLocalAccountAddress(address: AccountAddress): boolean {
@@ -357,43 +244,11 @@ export function useAccounts() {
     activeAccountGlobalIdx.value = 0;
   }
 
-  async function syncBackgroundEncryptionKey() {
-    await new Promise<void>((resolve) => {
-      const interval = setInterval(async () => {
-        const sessionEncryptionKey = await getSessionEncryptionKey();
-        if (sessionEncryptionKey) {
-          setEncryptionKey(sessionEncryptionKey);
-          clearInterval(interval);
-          resolve();
-        }
-      }, CHECK_FOR_SESSION_KEY_INTERVAL);
-    });
-  }
-
   (async () => {
-    if (!composableInitialized) {
-      composableInitialized = true;
-      const storedMnemonic = WalletStorage.get(STORAGE_KEYS.mnemonic);
-      if (
-        !encryptionKey.value
-        && !IS_MOBILE_APP
-        // If the mnemonic is stored but is not valid as plaintext
-        // it means that user is trying to access an existing & encrypted wallet
-        && storedMnemonic && !validateMnemonic(storedMnemonic)
-      ) {
-        await openPasswordLoginModal();
-      }
+    await watchUntilTruthy(isLoggedIn);
 
-      if (IS_OFFSCREEN_TAB && !encryptionKey.value) {
-        await syncBackgroundEncryptionKey();
-        return;
-      }
-
-      await watchUntilTruthy(isLoggedIn);
-
-      protocolLastActiveGlobalIdx
-        .value[activeAccount.value.protocol] = activeAccount.value.globalIdx;
-    }
+    protocolLastActiveGlobalIdx
+      .value[activeAccount.value.protocol] = activeAccount.value.globalIdx;
   })();
 
   return {
@@ -407,18 +262,12 @@ export function useAccounts() {
     activeAccount,
     activeAccountGlobalIdx,
     areAccountsRestored,
-    isMnemonicRestored,
     isLoggedIn,
     isActiveAccountAirGap,
-    mnemonic,
-    encryptionKey,
-    mnemonicSeed,
     protocolsInUse,
-    openPasswordLoginModal,
     discoverAccounts,
     isLocalAccountAddress,
     addRawAccount,
-    getEncryptionKey,
     getAccountByAddress,
     getAccountByGlobalIdx,
     getLastActiveProtocolAccount,
@@ -427,10 +276,6 @@ export function useAccounts() {
     setActiveAccountByGlobalIdx,
     setActiveAccountByProtocolAndIdx,
     setActiveAccountByProtocol,
-    setMnemonicAndInitializePassword,
-    setEncryptionKey,
-    updatePassword,
-    setGeneratedMnemonic,
     resetAccounts,
   };
-}
+});
