+++ src/composables/auth.ts
@@ -1,25 +1,86 @@
 import { BiometricAuth } from '@aparajita/capacitor-biometric-auth';
 import {
+  computed,
   readonly,
   ref,
   watch,
 } from 'vue';
+import { generateMnemonic, mnemonicToSeed } from '@aeternity/bip39';
 
 import { tg as t } from '@/popup/plugins/i18n';
-import { IS_MOBILE_APP, MODAL_ENABLE_BIOMETRIC_LOGIN, MODAL_BIOMETRIC_LOGIN } from '@/constants';
+import {
+  IS_EXTENSION,
+  IS_IOS,
+  IS_MOBILE_APP,
+  IS_OFFSCREEN_TAB,
+  MODAL_BIOMETRIC_LOGIN,
+  MODAL_ENABLE_BIOMETRIC_LOGIN,
+  MODAL_PASSWORD_LOGIN,
+  STORAGE_KEYS,
+} from '@/constants';
 import {
   authenticateWithPassword,
+  encrypt,
+  endSession,
+  generateEncryptionKey,
   getSessionEncryptionKey,
   sleep,
+  startSession,
   watchUntilTruthy,
 } from '@/utils';
+
+import { WalletStorage } from '@/lib/WalletStorage';
+import { SecureMobileStorage } from '@/lib/SecureMobileStorage';
+
+import migrateMnemonicVuexToComposable from '@/migrations/002-mnemonic-vuex-to-composable';
+import migrateMnemonicCordovaToIonic from '@/migrations/008-mnemonic-cordova-to-ionic';
+
 import { useUi } from './ui';
 import { useModals } from './modals';
-import { useAccounts } from './accounts';
 import { createCustomScopedComposable } from './composablesHelpers';
+import { useSecureStorageRef } from './secureStorageRef';
+import { useAeSdk } from './aeSdk';
+
+const CHECK_FOR_SESSION_KEY_INTERVAL = 5000;
 
 export const useAuth = createCustomScopedComposable(() => {
-  let composableInitialized = false;
+  const {
+    isBiometricLoginEnabled,
+    lastTimeAppWasActive,
+    secureLoginTimeout,
+    isAppActive,
+    setBiometricLoginEnabled,
+    setLoaderVisible,
+  } = useUi();
+  const { openModal } = useModals();
+
+  /** Common state for both biometric or password protection */
+  const isAuthenticated = ref(false);
+  const isAuthenticating = ref(false);
+  const isMnemonicRestored = ref(false);
+  const encryptionKey = ref<CryptoKey>();
+
+  const [mnemonicEncrypted, mnemonic] = useSecureStorageRef<string>(
+    '',
+    STORAGE_KEYS.mnemonic,
+    encryptionKey,
+    {
+      backgroundSync: true,
+      migrations: [
+        ...((IS_IOS && IS_MOBILE_APP) ? [migrateMnemonicCordovaToIonic] : []),
+        migrateMnemonicVuexToComposable,
+      ],
+      onRestored: async (val) => {
+        const hasStoredMnemonic = (
+          WalletStorage.get(STORAGE_KEYS.mnemonic)
+          || await SecureMobileStorage.get(STORAGE_KEYS.mnemonic)
+        );
+        isMnemonicRestored.value = !!val || !hasStoredMnemonic;
+      },
+    },
+  );
+
+  const mnemonicSeed = computed(() => mnemonic.value ? mnemonicToSeed(mnemonic.value) : null);
 
   const biometricAuth = {
     available: false,
@@ -27,30 +88,93 @@ export const useAuth = createCustomScopedComposable(() => {
     checked: false,
   };
 
-  /** Common state for both biometric or password protection */
-  const isAuthenticated = ref(false);
-  const isAuthenticating = ref(false);
-
-  const {
-    isBiometricLoginEnabled,
-    lastTimeAppWasActive,
-    secureLoginTimeout,
-    isAppActive,
-    setBiometricLoginEnabled,
-  } = useUi();
-  const { openModal } = useModals();
-  const {
-    isLoggedIn,
-    encryptionKey,
-    setEncryptionKey,
-    openPasswordLoginModal,
-  } = useAccounts();
-
   /**
    * TODO: Updating the authentication status from outside of the composable should be not allowed
    */
   function setAuthenticated(val: boolean) {
     isAuthenticated.value = val;
+  }
+
+  /**
+   * Setting/Resetting the password key logs the user in/out.
+   */
+  function setEncryptionKey(newEncryptionKey?: CryptoKey) {
+    encryptionKey.value = newEncryptionKey;
+    if (IS_EXTENSION) {
+      if (newEncryptionKey) {
+        startSession(newEncryptionKey, secureLoginTimeout.value);
+      } else {
+        endSession();
+      }
+    }
+  }
+
+  async function getEncryptionKey() {
+    return watchUntilTruthy(encryptionKey);
+  }
+
+  async function setPasswordAndEncryptMnemonic(newMnemonic: string, password: string) {
+    const newEncryptionKey = await generateEncryptionKey(password);
+    const mnemonicEncryptionResult = await encrypt(newEncryptionKey, newMnemonic);
+    mnemonicEncrypted.value = mnemonicEncryptionResult;
+    setEncryptionKey(newEncryptionKey);
+  }
+
+  async function openPasswordLoginModal() {
+    setLoaderVisible(true);
+    const sessionEncryptionKey = await getSessionEncryptionKey();
+    if (sessionEncryptionKey) {
+      setEncryptionKey(sessionEncryptionKey);
+      const { getAeSdk } = useAeSdk();
+      await getAeSdk();
+      setLoaderVisible(false);
+      return;
+    }
+    setLoaderVisible(false);
+
+    await openModal(MODAL_PASSWORD_LOGIN);
+    if (!encryptionKey.value) {
+      throw new Error('encryptionKey was not set after login.');
+    }
+  }
+
+  async function updatePassword(currentPassword: string, newPassword: string) {
+    const { decryptedMnemonic } = await authenticateWithPassword(currentPassword);
+    if (decryptedMnemonic) {
+      await setPasswordAndEncryptMnemonic(decryptedMnemonic, newPassword);
+    }
+  }
+
+  async function setMnemonicAndInitializePassword(newMnemonic: string, isRestored = false) {
+    if (!IS_MOBILE_APP) {
+      const { openSetPasswordModal } = useModals();
+
+      const password = await openSetPasswordModal(isRestored).catch(() => {
+        throw new Error('Password was not set.');
+      });
+
+      await setPasswordAndEncryptMnemonic(newMnemonic, password);
+    }
+    mnemonic.value = newMnemonic;
+  }
+
+  async function setGeneratedMnemonic() {
+    await setMnemonicAndInitializePassword(generateMnemonic()).catch(() => {
+      throw new Error('Mnemonic was not set.');
+    });
+  }
+
+  async function syncBackgroundEncryptionKey() {
+    await new Promise<void>((resolve) => {
+      const interval = setInterval(async () => {
+        const sessionEncryptionKey = await getSessionEncryptionKey();
+        if (sessionEncryptionKey) {
+          setEncryptionKey(sessionEncryptionKey);
+          clearInterval(interval);
+          resolve();
+        }
+      }, CHECK_FOR_SESSION_KEY_INTERVAL);
+    });
   }
 
   /**
@@ -69,6 +193,12 @@ export const useAuth = createCustomScopedComposable(() => {
       setBiometricLoginEnabled(false);
     }
     return biometricAuth.available;
+  }
+
+  async function openEnableBiometricLoginModal() {
+    if (await checkBiometricLoginAvailability()) {
+      openModal(MODAL_ENABLE_BIOMETRIC_LOGIN);
+    }
   }
 
   /**
@@ -105,22 +235,10 @@ export const useAuth = createCustomScopedComposable(() => {
   }
 
   /**
-   * Checks if the user should be kept logged in based on the current session.
-   */
-  async function shouldKeepExtensionLoggedIn() {
-    return !!(await getSessionEncryptionKey());
-  }
-
-  async function logout() {
-    setEncryptionKey(null);
-    isAuthenticated.value = false;
-  }
-
-  /**
    * Open biometric login or password login modal depending on the environment settings.
    */
   async function checkUserAuth(): Promise<any> {
-    if (!isAuthenticating.value && !isAuthenticated.value && isLoggedIn.value) {
+    if (!isAuthenticating.value && !isAuthenticated.value) {
       isAuthenticating.value = true;
       if (IS_MOBILE_APP) {
         if (isBiometricLoginEnabled.value && await checkBiometricLoginAvailability()) {
@@ -137,23 +255,31 @@ export const useAuth = createCustomScopedComposable(() => {
     }
   }
 
+  async function logout() {
+    setEncryptionKey(undefined);
+    isAuthenticated.value = false;
+  }
+
   async function lockWallet() {
     logout();
     checkUserAuth();
   }
 
-  async function openEnableBiometricLoginModal() {
-    if (await checkBiometricLoginAvailability()) {
-      openModal(MODAL_ENABLE_BIOMETRIC_LOGIN);
-    }
-  }
-
   (async () => {
-    if (composableInitialized) {
-      return;
+    checkBiometricLoginAvailability();
+
+    const encryptedMnemonicExists = !!(WalletStorage.get<string>(STORAGE_KEYS.mnemonic));
+    if (
+      !encryptionKey.value
+        && !IS_MOBILE_APP
+        && encryptedMnemonicExists
+    ) {
+      await checkUserAuth();
     }
-    composableInitialized = true;
-    await checkBiometricLoginAvailability();
+
+    if (IS_OFFSCREEN_TAB && !encryptionKey.value) {
+      await syncBackgroundEncryptionKey();
+    }
   })();
 
   watch(
@@ -167,7 +293,8 @@ export const useAuth = createCustomScopedComposable(() => {
         && isActive
         && !wasActive
       ) {
-        const keepExtensionLoggedIn = await shouldKeepExtensionLoggedIn();
+        // If session exists user needs to stay logged in
+        const keepExtensionLoggedIn = !!(await getSessionEncryptionKey());
         if (isAuthenticated.value && lastTimeAppWasActive.value) {
           const elapsedTime = Date.now() - lastTimeAppWasActive.value;
           if (elapsedTime > secureLoginTimeout.value && !keepExtensionLoggedIn) {
@@ -182,8 +309,17 @@ export const useAuth = createCustomScopedComposable(() => {
 
   return {
     isAuthenticated: readonly(isAuthenticated),
+    isMnemonicRestored,
+    mnemonic,
+    mnemonicEncrypted,
+    mnemonicSeed,
+    encryptionKey,
+    getEncryptionKey,
     setAuthenticated,
+    setGeneratedMnemonic,
+    setMnemonicAndInitializePassword,
     checkBiometricLoginAvailability,
+    updatePassword,
     authenticate,
     lockWallet,
     logout,
